should work? - moving the pc and sp to the machine itself as a 16 bit variables, so it   can now address upto 0xffff bytes, which is way too better than the   previous impl which can only address upto 256bytes.

+++ vm/src/memory.rs
@@ -2,18 +2,18 @@ use anyhow::Result;
 
 pub struct Memory {
     bytes: Vec<u8>,
-    size: u8,
+    size: u16,
 }
 
 impl Memory {
-    pub fn new(size: u8) -> Self {
+    pub fn new(size: u16) -> Self {
         Self {
             bytes: vec![0; size.into()],
             size,
         }
     }
 
-    pub fn read(&self, addr: u8) -> Result<u8> {
+    pub fn read(&self, addr: u16) -> Result<u8> {
         if addr < self.size {
             Ok(self.bytes[addr as usize])
         } else {
@@ -21,7 +21,7 @@ impl Memory {
         }
     }
 
-    pub fn write(&mut self, addr: u8, value: u8) -> Result<()> {
+    pub fn write(&mut self, addr: u16, value: u8) -> Result<()> {
         if addr < self.size {
             self.bytes[addr as usize] = value;
             Ok(())
@@ -32,7 +32,7 @@ impl Memory {
 
     pub fn load(&mut self, program_vec: &[u8]) -> Result<()> {
         for (index, byte) in program_vec.iter().enumerate() {
-            self.write(index as u8, *byte)?
+            self.write(index as u16, *byte)?
         }
         Ok(())
     }
